# Add experimental allowlist optimizations