--- a/config/flags/flags.go
+++ b/config/flags/flags.go
@@ -1,0 +1,6 @@
+package flags
+
+import "sync/atomic"
+
+// https://github.com/gitleaks/gitleaks/pull/1731
+var EnableExperimentalAllowlistOptimizations atomic.Bool