--- a/config/flags/flags.go
+++ b/config/flags/flags.go
@@ -1,6 +1,0 @@
-package flags
-
-import "sync/atomic"
-
-// https://github.com/gitleaks/gitleaks/pull/1731
-var EnableExperimentalAllowlistOptimizations atomic.Bool