--- a/detect/detect_test.go
+++ b/detect/detect_test.go
@@ -2275,6 +2275,9 @@ func TestDetectRuleAllowlist(t *testing.T) {
 let password = 'Summer2024!';`
 	for name, tc := range cases {
 		t.Run(name, func(t *testing.T) {
+			err := tc.allowlist.Validate()
+			require.NoError(t, err)
+
 			rule := config.Rule{
 				RuleID: "test-rule",
 				Regex:  regexp.MustCompile(`Summer2024!`),