### Description:
These are experimental changes to allowlist to improve efficiency.

Ideas include the following -- not all of them are implemented yet, or obviously better:

- [x] Using a map instead of slices for `Commits`
- [x] Collapsing `Regexes` into a single pattern
- [x] Collapsing `Paths` into a single pattern
- [x] Using ahocorasick for `Stopwords` instead of a slice
- [x] Creating benchmarks of varying sizes to get a more accurate portrayal of performance

**Commits**
```
# Before
BenchmarkCommitAllowed
BenchmarkCommitAllowed-4      	 3539479	       335.0 ns/op
BenchmarkCommitNotAllowed
BenchmarkCommitNotAllowed-4   	 2950525	       403.2 ns/op

# After
BenchmarkCommitAllowed
BenchmarkCommitAllowed-4      	16353297	        79.87 ns/op
BenchmarkCommitNotAllowed
BenchmarkCommitNotAllowed-4   	17393956	        69.91 ns/op
```

**Paths**
```
# Before
BenchmarkPathAllowed
BenchmarkPathAllowed-4        	   68570	     15523 ns/op
BenchmarkPathNotAllowed
BenchmarkPathNotAllowed-4     	   41542	     27972 ns/op

# After
BenchmarkPathAllowed
BenchmarkPathAllowed-4        	  922988	      1256 ns/op
BenchmarkPathNotAllowed
BenchmarkPathNotAllowed-4     	  767101	      1474 ns/op
```

**Regexes**
```
# Before
BenchmarkRegexAllowed
BenchmarkRegexAllowed-4       	  100308	     10873 ns/op
BenchmarkRegexNotAllowed
BenchmarkRegexNotAllowed-4    	   38066	     31017 ns/op

# After
BenchmarkRegexAllowed
BenchmarkRegexAllowed-4       	 1243234	       947.4 ns/op
BenchmarkRegexNotAllowed
BenchmarkRegexNotAllowed-4    	  835924	      1572 ns/op
```

cc @zricethezav @bplaxco 

### Potential Drawbacks

Off the top of my head, potential drawbacks could be:

- Slower startup speed to build an ahocorasick trie for every allowlist
- Faster performance for larger allowlists, at the cost of slower or equivalent performance for smaller allowlists
- Changes the semantics of the `Allowlist` struct, could mess with how some people use Gitleaks as a library

### Checklist:

* [x] Does your PR pass tests?
* [x] Have you written new tests for your changes?
* [x] Have you lint your code locally prior to submission?