@bplaxco 2025-01-28
> Allowlist struct, could mess with how some people use Gitleaks as a library

Yeah anything that we can do to keep state/caches on the detector would be handy for my use case but I can always tweak things. I load the config once and then have a process that continually scans using gitleaks as a lib. I do create a fresh detector each scan though.