@rgmz 2025-01-29
> For my own understanding—apologies if I’m missing something about this codebase or the scanning approach—how do these optimizations affect hot paths? Does a typical scan run through all the optimized sections, or are there specific hotspots?

I suspect it runs through most, however, I don't have any data to support that. It would be interesting to gather.

Some users maintain their own config, [others (likely most) use the default config](https://open.substack.com/pub/lookingatcomputer/p/regex-is-almost-all-you-need?selection=8ddbbd08-f43c-45d6-b099-f0b48846f921). Moreover, the same config can have wildly different performance depending on what it's scanning. Every fragment is evaluated against [the global `[allowlist]`](https://github.com/gitleaks/gitleaks/blob/3188ad60055da792f5dd2bba21a88f4d5be80f96/config/gitleaks.toml#L15-L57). Every finding is evaluated against the rule's allowlist, if present (e.g., [`curl-auth-user`](https://github.com/gitleaks/gitleaks/blob/3188ad60055da792f5dd2bba21a88f4d5be80f96/config/gitleaks.toml#L299)), but no rule is guaranteed to hit. [`generic-api-key`](https://github.com/gitleaks/gitleaks/blob/3188ad60055da792f5dd2bba21a88f4d5be80f96/config/gitleaks.toml#L566) is most likely to hit and also has the most complicated allowlist.

> Also, out of curiosity, do you have a CPU or trace profile that guided these optimizations?

I do not. Are there any resources you recommend to get started with that?