@rgmz 2025-03-02
On a large (15GB) repository, this reduced the scanning time from ~46m to ~32m.

**Before**
```
(1) 2:53PM INF scanned ~15450035936 bytes (15.45 GB) in 46m53s
...
(2) 7:27PM INF scanned ~15450035936 bytes (15.45 GB) in 46m7s
```

**After**
```
(1) 5:25PM INF scanned ~15450035936 bytes (15.45 GB) in 32m58s
...
(2) 6:15PM INF scanned ~15450035936 bytes (15.45 GB) in 32m41s
```