@zricethezav 2025-06-07
> That sounds good to me! What do you think about https://github.com/gitleaks/gitleaks/pull/1731#discussion_r2127816521?

Sure, that sounds reasonable 👍🏻