@rgmz 2025-06-07
I just realized the current implementation won't work for extended configs. Gotta run out now, will try to fix it later.